Switch ament_package to using importlib.
As part of the switch away from pkg_resources (since it is
slow), we also need to fix up ament_package.  ament_package was using
'resource_filename' from pkg_resources.  The equivalent in
importlib is importlib.resources, but there is a catch.
As of Python 3.8, importlib.resources does not support getting
data from subdirectories.  The workaround is to make each
level of the directory structure a python module, so that the
module name can be looked up.  Python 3.9 will have support for
directories for resources, but that won't help on Ubuntu Focal.
More information is in https://gitlab.com/python-devs/importlib_resources/issues/58

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>