Simplify removing leading and trailing separators (#152)
Previously, we checked if the leading or trailing character was a colon
and then used a wildcard to remove it.

It is simpler to just remove a leading or trailing colon. This has the
added benefit of only using shell built-in functions.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>
(cherry picked from commit 9b09d9471f8bf77d5c847a6cd973da79d18d2e11)